Move the ar test package into game package and add new tests
