+++ mobile/src/androidTest/java/ch/epfl/sdp/mobile/test/ui/game/ar/ArScreenKtTest.kt
@@ -1,4 +1,4 @@
-package ch.epfl.sdp.mobile.test.ui.ar
+package ch.epfl.sdp.mobile.test.ui.game.ar
 
 import android.Manifest
 import android.app.Activity
